pyproject.toml
[tool.poetry]
name = "epub_with_python"
version = "0.1.0"
description = ""
authors = ["cu39 <cu393uc@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
invoke = "^1.7.0"
PyYAML = "^6.0"
Jinja2 = "^3.1.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
